Fix broken tests of initial cleaning helper functions.
